Remove form-response POST retry
This is batching a bunch of `POST`s for form drafts, and the first one sent may “win” for the user.

This was added prior to form drafts and `PATCH`ing to the draft.  So while this is broken it’s also incomplete.  For now we can rely on the local cache of the form draft when offline.