Normalize api query between the list and single action/flow gets
Previously we were relying on default API settings on the action/flow list GETs.  fields only worked with includes, but they were auto-applied includes.

If we’re explicit, then it is the same request for the list and the request we need to make to add a newly notified entity.